chore: change typings to local
this changes typings to local. it's worth noting, there is not exactly a benefit to this, as this is a template, and changes to the NPM package would not be reflected in people's clones of this template as a result.
however, the NPM packages have not been updated in a while and are now outdated/inaccurate, and in the interest of keeping things carefree and adding less stress to your maintenance over a lightweight template, i think this would help people as if they wanted updated typings, they could simply fetch the latest from the template.