feat(craftos): add SlotDetail typings
This commit adds an explicitly named type for the return of the `list()` function on an inventory peripheral.
This allows for people to create their own typings based off of the slot detail or reference the type as part of their own code.

This also contains one type fix, changing the `count` field on `ItemDetail` to a number.